fix type bug

+++ train.py
@@ -28,7 +28,7 @@ def forward_pass(x, y_reg, y_cor, y_dontcare=None):
         # downsample GT y_cor s.t. compute loss at low resolution
         B, C, W = y_cor.shape
         y_cor = y_cor.reshape([B, C, W//args.y_step, args.y_step] )
-        y_cor = y_cor.sum(3)
+        y_cor = y_cor.sum(3).float()
         if args.ori_res_loss:
             # upsample Pred y_reg_ s.t. compute loss at full resolution
             ori_w = y_reg.shape[2]
